+++ app/update_catalog_handler.go
@@ -7,8 +7,8 @@ import (
 	"bitbucket.org/andyfusniakteam/ecom-api-go/utils/nestedset"
 )
 
-// ReplaceCatalogHandler creates an HTTP handler that creates a new customer address record.
-func (a *App) ReplaceCatalogHandler() http.HandlerFunc {
+// UpdateCatalogHandler creates an HTTP handler that updates the catalog.
+func (a *App) UpdateCatalogHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		catalog := nestedset.Node{}
 		err := json.NewDecoder(r.Body).Decode(&catalog)
@@ -24,7 +24,7 @@ func (a *App) ReplaceCatalogHandler() http.HandlerFunc {
 			return
 		}
 		defer r.Body.Close()
-		a.Service.ReplaceCatalog(r.Context(), &catalog)
+		a.Service.UpdateCatalog(r.Context(), &catalog)
 		w.Header().Del("Content-Type")
 		w.WriteHeader(http.StatusNoContent) // 204 No Content
 	}